@@ -8,26 +8,29 @@
     "production": "cross-env NODE_ENV=production webpack --progress --config=webpack.config.js"
   },
   "dependencies": {
-    "axios": "^0.21.4",
-    "core-js": "^3.17.3",
+    "@vue/compat": "^3.2.21",
+    "axios": "^0.24.0",
+    "core-js": "^3.19.1",
     "es6-promise": "^4.2.8",
-    "vue": "^2.6.14",
-    "vuex": "^3.6.2"
+    "mitt": "^3.0.0",
+    "vue": "^3.2.21",
+    "vuex": "^4.0.2"
   },
   "devDependencies": {
-    "babel-eslint": "^10.1.0",
-    "babel-loader": "^8.2.2",
+    "@babel/eslint-parser": "^7.16.0",
+    "@babel/core": "^7.16.0",
+    "@babel/preset-env": "^7.16.0",
+    "babel-loader": "^8.2.3",
     "craftcms-sass": "^3.5.7",
     "cross-env": "^7.0.3",
-    "css-loader": "^6.2.0",
-    "eslint": "^7.32.0",
-    "eslint-plugin-vue": "^7.17.0",
-    "sass": "^1.40.1",
-    "sass-loader": "^12.1.0",
-    "vue-loader": "^15.9.8",
+    "css-loader": "^6.5.0",
+    "eslint": "^8.1.0",
+    "eslint-plugin-vue": "^8.0.3",
+    "sass": "^1.43.4",
+    "sass-loader": "^12.3.0",
+    "vue-loader": "^16.8.2",
     "vue-style-loader": "^4.1.3",
-    "vue-template-compiler": "^2.6.14",
-    "webpack": "^5.52.1",
-    "webpack-cli": "^4.8.0"
+    "webpack": "^5.61.0",
+    "webpack-cli": "^4.9.1"
   }
 }
