@@ -1,66 +1,51 @@
 import 'es6-promise/auto';
-
-import Vue from 'vue';
-
-import { EventBus } from './EventBus.js';
-
-import TranslationsStore from './store/TranslationsStore';
-
+import {createApp} from 'vue';
+import {store} from './store/TranslationsStore';
 import SaveTranslations from './components/SaveTranslations.vue';
 import TranslationsList from './components/TranslationsList.vue';
 import TranslationsToolbar from './components/TranslationsToolbar.vue';
 import TranslationsMenu from './components/TranslationsMenu.vue';
 import TranslationsFooter from './components/TranslationsFooter.vue';
-import { mapState } from 'vuex';
+import {mapState} from 'vuex';
+import mitt from 'mitt';
 
-Vue.prototype.$csrfTokenName = window.csrfTokenName;
-Vue.prototype.$csrfTokenValue = window.csrfTokenValue;
-Vue.prototype.$craft = window.Craft;
+const emitter = mitt();
 
-new Vue({
-  el: '#main',
-  store: TranslationsStore,
-  components: {
-    SaveTranslations,
-    TranslationsList,
-    TranslationsToolbar,
-    TranslationsMenu,
-    TranslationsFooter
-  },
+const app = createApp({
   computed: {
     ...mapState({
       category: state => state.category,
     })
   },
-  created () {
-    EventBus.$on('translations-saved', () => {
+  created() {
+    this.emitter.on('translations-saved', () => {
       this.$craft.cp.displayNotice(this.$craft.t('translations-admin', 'Translations saved'));
     });
-    EventBus.$on('translations-saved-error', () => {
+    this.emitter.on('translations-saved-error', () => {
       this.$craft.cp.displayError(this.$craft.t('translations-admin', 'Translations not saved'));
     });
-    EventBus.$on('translation-added', () => {
+    this.emitter.on('translation-added', () => {
       this.$craft.cp.displayNotice(this.$craft.t('translations-admin', 'Translation added'));
     });
-    EventBus.$on('translation-added-error', () => {
+    this.emitter.on('translation-added-error', () => {
       this.$craft.cp.displayError(this.$craft.t('translations-admin', 'Translation not added'));
     });
-    EventBus.$on('translation-deleted', () => {
+    this.emitter.on('translation-deleted', () => {
       this.$craft.cp.displayNotice(this.$craft.t('translations-admin', 'Translation deleted'));
     });
-    EventBus.$on('translation-deleted-error', () => {
+    this.emitter.on('translation-deleted-error', () => {
       this.$craft.cp.displayError(this.$craft.t('translations-admin', 'Translation not deleted'));
     });
   },
-  mounted () {
+  mounted() {
     // Redo Jquery selectors after Vue has mounted
-    window.Craft.cp.$headerContainer = window.$('#header-container');
-    window.Craft.cp.$header = window.$('#header');
-    window.Craft.cp.$mainContent = window.$('#main-content');
-    window.Craft.cp.$details = window.$('#details');
-    window.Craft.cp.$sidebarContainer = window.$('#sidebar-container');
-    window.Craft.cp.$sidebar = window.$('#sidebar');
-    window.Craft.cp.$contentContainer = window.$('#content-container');
+    this.$craft.cp.$headerContainer = this.$jquery('#header-container');
+    this.$craft.cp.$header = this.$jquery('#header');
+    this.$craft.cp.$mainContent = this.$jquery('#main-content');
+    this.$craft.cp.$details = this.$jquery('#details');
+    this.$craft.cp.$sidebarContainer = this.$jquery('#sidebar-container');
+    this.$craft.cp.$sidebar = this.$jquery('#sidebar');
+    this.$craft.cp.$contentContainer = this.$jquery('#content-container');
 
     if (document.querySelector('#sidebar-toggle')) {
       document.querySelector('#sidebar-toggle').addEventListener('click', () => {
@@ -69,7 +54,7 @@ new Vue({
     }
   },
   watch: {
-    category () {
+    category() {
       if (document.querySelector('#selected-sidebar-item-label')) {
         document.body.classList.remove('showing-sidebar');
         document.querySelector('#selected-sidebar-item-label').innerHTML = this.category;
@@ -77,3 +62,20 @@ new Vue({
     }
   },
 });
+
+app
+  .component('SaveTranslations', SaveTranslations)
+  .component('TranslationsList', TranslationsList)
+  .component('TranslationsToolbar', TranslationsToolbar)
+  .component('TranslationsMenu', TranslationsMenu)
+  .component('TranslationsFooter', TranslationsFooter);
+
+app.use(store);
+
+app.config.globalProperties.$csrfTokenName = window.csrfTokenName;
+app.config.globalProperties.$csrfTokenValue = window.csrfTokenValue;
+app.config.globalProperties.$jquery = window.$;
+app.config.globalProperties.$craft = window.Craft;
+app.config.globalProperties.emitter = emitter;
+
+app.mount('#main');
