@@ -1,18 +1,12 @@
 module.exports = {
-  root: true,
-  env: {
-    node: true
-  },
-  'extends': [
-    'plugin:vue/essential',
-    'eslint:recommended'
+  extends: [
+    'eslint:recommended',
+    'plugin:vue/vue3-essential'
   ],
   rules: {
     'semi': [2, 'always'],
     'no-console': process.env.NODE_ENV === 'production' ? 'error' : 'off',
-    'no-debugger': process.env.NODE_ENV === 'production' ? 'error' : 'off'
+    'no-debugger': process.env.NODE_ENV === 'production' ? 'error' : 'off',
+    'no-undef': 'off'
   },
-  parserOptions: {
-    parser: 'babel-eslint'
-  }
 };
